refactor and relocate CORS config
provides mapping for /auth and /api; their allowances are identical for now
