+++ BarkBook-API/src/main/java/org/teampraxis/BarkBook_API/BarkBookApiApplication.java
@@ -2,9 +2,6 @@ package org.teampraxis.BarkBook_API;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
 public class BarkBookApiApplication {
@@ -13,18 +10,4 @@ public class BarkBookApiApplication {
 		SpringApplication.run(BarkBookApiApplication.class, args);
 	}
 
-	// cors config
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/api/**")
-						.allowedOrigins("http://localhost:5173/")
-						.allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-						.allowedHeaders("*")
-						.allowCredentials(true);
-			}
-		};
-	}
 }
